Refactored package organization.
