--- src/easy/multithreading/system/ISystemWrapper.java
@@ -1,5 +1,0 @@
-package easy.multithreading.system;
-
-public interface ISystemWrapper {
-    void print(String thingToPrint);
-}
