--- src/easy/multithreading/system/SystemWrapper.java
@@ -1,13 +1,0 @@
-package easy.multithreading.system;
-
-class SystemWrapper implements ISystemWrapper {
-
-    public SystemWrapper() {
-    }
-
-    @Override
-    public void print(String thingToPrint) {
-        System.out.print(thingToPrint);
-    }
-
-}
