+++ src/printinorder/multithreading/hard/ConcurrentThreadExecutorWithSingleton.java
@@ -1,6 +1,6 @@
-package hard.multithreading;
+package printinorder.multithreading.hard;
 
-import hard.multithreading.system.ISystemWrapper;
+import printinorder.multithreading.system.ISystemWrapper;
 
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
