+++ src/printinorder/multithreading/system/ISystemWrapper.java
@@ -1,4 +1,4 @@
-package hard.multithreading.system;
+package printinorder.multithreading.system;
 
 public interface ISystemWrapper {
     void print(String thingToPrint);
