+++ src/printinorder/multithreading/system/SystemWrapper.java
@@ -1,4 +1,4 @@
-package hard.multithreading.system;
+package printinorder.multithreading.system;
 
 class SystemWrapper implements ISystemWrapper {
